Version 2.1.10
- update of conditional compile conditions to macros of actual STM32 package

+++ src/GxIO/GxIO_SPI/GxIO_SPI.cpp
@@ -493,7 +493,7 @@ void GxIO_SPI_USING_TRANSACTION::setBackLight(bool lit)
 
 #else
 
-GxIO_SPI_USING_TRANSACTION::GxIO_SPI_USING_TRANSACTION(SPIClass& spi, int8_t cs, int8_t dc, int8_t rst, int8_t bl)
+GxIO_SPI_USING_TRANSACTION::GxIO_SPI_USING_TRANSACTION(SPIClass& spi, int16_t cs, int16_t dc, int16_t rst, int16_t bl)
   : IOSPI(spi), settings(GxIO_SPI_defaultFrequency, MSBFIRST, SPI_MODE0)
 {
   _cs   = cs;
